+++ contracts/src/Types.sol
@@ -54,12 +54,12 @@ library Types {
     }
 
     // events for key-gen
-    event SecretGenRound1(uint256 indexed oprfKeyId, uint256 threshold);
-    event SecretGenRound2(uint256 indexed oprfKeyId);
-    event SecretGenRound3(uint256 indexed oprfKeyId);
-    event SecretGenFinalize(uint256 indexed oprfKeyId);
+    event SecretGenRound1(uint160 indexed oprfKeyId, uint256 threshold);
+    event SecretGenRound2(uint160 indexed oprfKeyId);
+    event SecretGenRound3(uint160 indexed oprfKeyId);
+    event SecretGenFinalize(uint160 indexed oprfKeyId);
     // event to delete created key
-    event KeyDeletion(uint256 indexed oprfKeyId);
+    event KeyDeletion(uint160 indexed oprfKeyId);
     // admin events
     event KeyGenAdminRevoked(address indexed admin);
     event KeyGenAdminRegistered(address indexed admin);
